+++ CODEFORCES/1500/1603B.cpp
@@ -1,0 +1,16 @@
+#include <bits/stdc++.h>
+using namespace std;
+int main(){
+    int T; cin >> T;
+    while(T--){
+        long long x, y;
+        cin >> x >> y;
+        if(x==y){
+            cout << x << endl;
+        }else if(x > y){
+            cout << x+y << endl;
+        }else{
+            cout << y - y%x/2 << endl;
+        }
+    }
+}