+++ USACO/Contests/2023-silver-preidolia.cpp
@@ -1,0 +1,17 @@
+#include <bits/stdc++.h>
+using namespace std;
+int main(){
+    int T; cin >> T;
+    while(T--){
+        string s; cin >> s;
+        int n = s.size();
+        int s, e;
+        string tar = "bessie";
+        vector<int> f(6);
+        for(int i = 0; i < n; i++){
+            if(s[i] == tar[0]){
+
+            }
+        }
+    }
+}